Generated code should only use the helper methods that are actually used.
This is determined by simple string checking and in the event of incorrect usage detection, will over-include things which is objectively better than what's happening now (everything is included).

The Java helper methods are nearly 400 lines long and was creating lots of waste.

Additionally, this allows a much more extensive set of core functions without worrying if it will bloat the output.