Add a core function called ExtensibleCallbackInvoke
The exported code has a function called ExtensibleCallbackRegister which allows you to register arbitrary functions that take in an object array and return an object to a given string name. Then the pastel code can invoke the function if it's registered (or return null if it's not). The current extensibility mechanism is somewhat broken and Crayon-specific as it relies on the existence of a legacy class called TranslationHelper to exist in order for the generated Pastel code to compile, which still dates back to Pastel and Crayon's tightly coupled days. Ideally that can be removed in favor of this, particularlly for Crayon's CNI registration methods while being generic enough for other uses.